Integrate store. Reorganize hierarchy
