+++ js/components/HomeScene/Scene/Rig/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./Rig";
